% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_function.R
\name{aggregate_time_series}
\alias{aggregate_time_series}
\title{A function that aggregates time series data}
\usage{
aggregate_time_series(data, ORDERDATE, SALES, time_unit = "month")
}
\arguments{
\item{data}{A tibble format data, including the following columns:
\itemize{
\item ORDERDATE: A column in POSIXct format, e.g., "2020-02-24", representing the order dates.
\item SALES: A column recording the sales price.
}}

\item{ORDERDATE}{A column in POSIXct format, e.g., "2020-02-24", representing the order dates.}

\item{SALES}{A column recording the sales price.}

\item{time_unit}{A character string specifying the time unit to aggregate the data (default: "month").}
}
\value{
A tibble with the aggregated data, including the date and total sales.
}
\description{
This function aggregates the given data based on the specified time unit.
time_unit value includes "day", "week", "month", "quarter", "year".
Then it calculates the total sales for each time unit and generates the time series info.
}
\examples{
# Example data
data <- tibble::tribble(
  ~ORDERDATE, ~SALES,
  "2020-02-24", 2871,
  "2020-05-07", 2766,
  "2020-08-30", 3884,
  # More rows...
)
data$ORDERDATE=as.POSIXct(data$ORDERDATE)
aggregated_data <- aggregate_time_series(data,
                                         time_unit = "month")
}
\author{
Ting Wei
}
