% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_function.R
\name{generate_forecast}
\alias{generate_forecast}
\title{A function that generates forecasts for the time series data}
\usage{
generate_forecast(data, n_future = 12, seed = NULL)
}
\arguments{
\item{data}{A data frame containing the time series data.}

\item{n_future}{An integer specifying the number of future periods to forecast (default: 12 and min:1).}

\item{seed}{An optional seed for reproducibility (default: NULL).}
}
\value{
A data frame with the actual and predicted values for the time series data.
}
\description{
This function generates forecasts for the future time periods based on the given data.
}
