% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_function.R
\name{plot_forecast}
\alias{plot_forecast}
\title{A function that plots the forecasts for the time series data}
\usage{
plot_forecast(data)
}
\arguments{
\item{data}{A data frame containing the time series data and forecasts.}
}
\value{
A plot of the actual and predicted values.
}
\description{
This function plots the actual and predicted values for the time series data.
}
