% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_function.R
\name{plot_time_series}
\alias{plot_time_series}
\title{aggregated_data <- aggregate_time_series(data, time_unit = "month")
A function that plots the time series data}
\usage{
plot_time_series(data)
}
\arguments{
\item{data}{A data frame containing the time series data.}
}
\value{
A plot of the time series data.
}
\description{
This function plots the time series data, showing the total sales over time.
}
